refactor: extract initial data fetching to a hook to improve readability
